create env file and read config from it
