+++ app/.env.template
@@ -1,0 +1,5 @@
+APP_CONFIG__DATABASE__URL=postgresql+asyncpg://user:pswd@host:port/name
+APP_CONFIG__DATABASE__ECHO=1
+APP_CONFIG__DATABASE__ECHO_POOL=0
+APP_CONFIG__DATABASE__POOL_SIZE=50
+APP_CONFIG__DATABASE__MAX_OVERFLOW=10
