--- app/config/__init__.py
@@ -1,1 +1,0 @@
-from .conf import config
