--- app/config/conf.py
@@ -1,26 +1,0 @@
-from pydantic import BaseModel, PostgresDsn
-from pydantic_settings import BaseSettings
-
-class RunConfig(BaseModel):
-    host: str = "0.0.0.0"
-    port: int = 7654
-
-
-class ApiPrefix(BaseModel):
-    prefix: str = "/api"
-
-
-class DBConfig(BaseModel):
-    url: PostgresDsn
-    echo: bool = True
-    echo_pool: bool = False
-    pool_size: int = 50,
-    max_overflow: int = 10
-
-class Config(BaseSettings):
-    run: RunConfig = RunConfig()
-    api: ApiPrefix = ApiPrefix()
-    db: DBConfig
-
-
-config = Config()
